+++ packages/browser-wallet/src/popup/pages/SendTransaction/SendTransaction.tsx
@@ -14,7 +14,7 @@ import {
     getAccountTransactionHash,
     SchemaVersion,
 } from '@concordium/web-sdk';
-import { jsonRpcUrlAtom, credentialsAtom } from '@popup/store/settings';
+import { credentialsAtom, networkConfigurationAtom } from '@popup/store/settings';
 import DisplayUpdateContract from './displayTransaction/DisplayUpdateContract';
 import DisplayInitContract from './displayTransaction/DisplayInitContract';
 import DisplaySimpleTransfer from './displayTransaction/DisplaySimpleTransfer';
@@ -43,7 +43,7 @@ export default function SendTransaction({ onSubmit, onReject }: Props) {
     const { t } = useTranslation('sendTransaction');
     const [error, setError] = useState<string>();
     const creds = useAtomValue(credentialsAtom);
-    const url = useAtomValue(jsonRpcUrlAtom);
+    const { jsonRpcUrl } = useAtomValue(networkConfigurationAtom);
     const { withClose, onClose } = useContext(fullscreenPromptContext);
 
     const { accountAddress } = state.payload;
@@ -62,7 +62,7 @@ export default function SendTransaction({ onSubmit, onReject }: Props) {
     useEffect(() => onClose(onReject), [onClose, onReject]);
 
     const sendTransaction = useCallback(async () => {
-        if (!url || !accountAddress) {
+        if (!jsonRpcUrl || !accountAddress) {
             throw new Error('Missing url for JsonRpc or account address');
         }
         const key = creds.find((c) => c.address === accountAddress)?.key;
@@ -71,7 +71,7 @@ export default function SendTransaction({ onSubmit, onReject }: Props) {
         }
 
         // TODO: Maybe we should not create the client for each transaction sent
-        const client = new JsonRpcClient(new HttpProvider(url));
+        const client = new JsonRpcClient(new HttpProvider(jsonRpcUrl));
         const sender = new AccountAddress(accountAddress);
         const nonce = await client.getNextAccountNonce(sender);