+++ packages/browser-wallet/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 -   Login screen.
 
+### Changed
+
+-   The network connections page has been updated with new UI.
+
 ## 0.3.0 2022-8-18
 
 ### Added